Updated Matrix and Vector tempPtrs to have the appropriate size.

+++ neo/idlib/math/MatX.cpp
@@ -36,7 +36,7 @@ If you have questions concerning this license or the applicable additional terms
 //===============================================================
 
 float	idMatX::temp[MATX_MAX_TEMP+4];
-float *	idMatX::tempPtr = (float *) ( ( (int) idMatX::temp + 15 ) & ~15 );
+float *	idMatX::tempPtr = (float *) ( ( (uintptr_t) idMatX::temp + 15 ) & ~15 );
 int		idMatX::tempIndex = 0;
